+++ include/tig/idxtable.h
@@ -38,7 +38,7 @@ typedef bool(TigIdxTableEnumerateCallback)(void* value, int key, void* context);
 void tig_idxtable_init(TigIdxTable* idxtable, int value_size);
 void tig_idxtable_exit(TigIdxTable* idxtable);
 void tig_idxtable_copy(TigIdxTable* dst, TigIdxTable* src);
-int tig_idxtable_write(TigIdxTable* idxtable, TigFile* stream);
+size_t tig_idxtable_write(TigIdxTable* idxtable, TigFile* stream);
 bool tig_idxtable_read(TigIdxTable* idxtable, TigFile* stream);
 bool tig_idxtable_enumerate(TigIdxTable* idxtable, TigIdxTableEnumerateCallback* callback, void* context);
 void tig_idxtable_set(TigIdxTable* idxtable, int key, void* value);
